core: comma separated list of values in search
Add metadata in search engine that enables to support column values that
actually includes comma separated lists of values
For example, if a ISO storage is attached to 2 DCs : dc1 and dc2 then in
the storage_domains_for_search view the column datacenter value is
"dc1,dc2" , current search syntax will search for 'dc1' or 'dc2' from
the tree view and will give no results.
The same applies for a disk that is shared by 2 VMs vm2 and vm2 when searching from
the disk tab for vm_names with 'vm1' or 'vm2'
This patch adds an option to add metadata that enabled to define per
table/view a list of columns that have actually comma separated values
as "dc1,dc2" or "vm1,vm2" in the examples above.
Once this metadata is added, the search engine will build the proper SQL
to match any of the comma separated values.

Change-Id: Ie0fa048b1d12392f1c1ea81ed720be6c4078f7d0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1320774
Signed-off-by: emesika <emesika@redhat.com>