core: Connect all hosts on adding storage pool.
Once we create a storage pool with multiple hosts, we connect the first
host to all the storage domains and create a storage pool.
After the create succeeded, the flow continues, and tries to fetch
all the unregistered disks before all the hosts in the Data Center
are connected to all the storage domains.

As part of the query of getting the unregistered disks, the engine uses
the getImagesList query.
This vds command uses IrsProxyData which executes
ConnectStoragePoolVDSCommand, and the host that execute this method uses
a random host which might not be connected.

To avoid this problem, once we have a host which successfully created a
storage pool we keep connect all the rest of the storage domains since
they are already being connected when getting activated.

Change-Id: I57ca7d7495410da6caada1c08d2ed95d467a464c
Bug-Url: https://bugzilla.redhat.com/1344314
Signed-off-by: Maor Lipchuk <maorlipchuk@gmail.com>