restapi: Document the rename of "Slaves" to "HostNics"
In version 4 of the API the names of the XML complex types are
systematically generated from the names of the types of the model. For
example, the name for the XML complex type "HostNic" is "HostNics". But
in version 3 of the API this specific type was named differently,
because the XML schema was written manually. This patch updates the
documentation to include this type name change.

Change-Id: Ia231664e171ed4390557f7bf5662c12ec548ca0f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>