webadmin: Make DiskModel's default interface not abstract.
Disk's interface has nothing to do with the kind of Model inherited
from AbstractDiskModel (e.g NewDiskModel, EditDiskModel, ReadOnlyDiskModel).

If a disk is attached to a VM, the model should retrieve the right interface.
otherwise, it should just take a default one.

The current implementation also caused an error in ReadOnlyDiskModel,
when resuming an upload. The model tried to retrieve the selected
interface, but since the disk was not attached to a VM, an error was
thrown.

Change-Id: I0e85d980e130a973a865a54eb82758b8f8e30dfb
Bug-Url: https://bugzilla.redhat.com/1344367
Signed-off-by: Amit Aviram <aaviram@redhat.com>
(cherry picked from commit c61b9ba573ccd0f2b6bf709a60627136d89fc5de)