restapi: Don't add name to network filter links
Currently the links to network filters included in the representation of
VNIC profiles include also the name of the network filters. This is
contrary to the common practice in the API, where links only have the
"id" and "href" attributes. Other attributes or elements are allowed
only when required for backwards compatbility. This patch removes the
code that adds the network filter name.

Change-Id: I3379bbf794cbec22a059544bf9684fd160ade2d0
Bug-Url: https://bugzilla.redhat.com/1347931
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>