webadmin: fix UploadDiskModel validation and flush.
Currently, the validation of the model gets a parameter that determines
if the object's diskModel validation will be called. When starting the
upload, this function is called 2 times, once when flushing, then before
starting the upload.

In case that the diskModel validation failed, it didn't flush, causing
an NPE when calling onUpload() without filling the needed fields- as the
diskModel validation is not called before flushing.

This patch makes the validation always call to the diskModel validation,
and makes flush() to return false if the model wasn't able to flush.

Change-Id: I2fc1942ffffd60254d1344fce582afbb6562c697
Bug-Url: https://bugzilla.redhat.com/1328887
Signed-off-by: Amit Aviram <aaviram@redhat.com>