core: Allow user to create a vm from blank template on local storage
When user creates a new vm from blank template on local storage the vm
is pinned to the one host that exists on the storage.
Since the blank template is not pinned to a specific host the
authorization for that action fails (the dedicated host of the vm is
compared to the dedicated host of the template and since only admin is
allowed to change hosts and the template is not pinned an exception is
thrown).
The fix is to remove the comparison of hosts in case the vm is created
from a default template.

Change-Id: I5f30844e3f992e4a1454e0ddeab6ae4d13591997
Bug-Url: https://bugzilla.redhat.com/1350223
Signed-off-by: Jenny Tokar <jtokar@redhat.com>
(cherry picked from commit 4522754)
Signed-off-by: Jenny Tokar <jtokar@redhat.com>