restapi: Populate V3 snapshot type correctly
In version 3 of the API the VmBase#type attribute had two different
meanings, depending on the context: the type of virtual machine (desktop
of server) when used with virtual machines or templates, and the type of
snapshot (regular, active, etc) when sued with snapshots. In version 4
of the API this isn't possible because each attribute has attribute has
exactly one type. To handle that a new `snapshot_type` attribute was
added to the snapshot entity. But the version 3 compatibility layer
didn't take this into account, and as a result the old `type` attribute
wasn't populated. This patch fixes that, so the old `type` attribute is
populated correctly.

Change-Id: I8b50b994533135fbc0a5d924a3df958b4b6cb584
Bug-Url: https://bugzilla.redhat.com/1348138
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 977aa6a2222d9166bad240d9cef69fe062026043)