core: acquire vm lock on RestoreStatelessVm
RestoreStatelessVm calls two commands:
1. UpdateVmVersion
2. RestoreAllSnapshots

Both used to take exclusive lock on the VM.
Change I870ef6 modified UpdateVmVersion so it will not
acquire the VM lock when it is called by RestoreStatelessVm
but that's because we assume RestoreStatelessVm locked the
VM which is not always the case.

So this patch changes RestoreStatelessVm to lock the VM.
It is safe to do since UpdateVmVersion and RestoreAllSnapshots
used to do it later in the control flow as well. This lock is
passed to RestoreAllSnapshot so it doesn't try to lock the VM
as well.

Change-Id: I3e52fafa8fcb1976c309e5ea79227e9dc1a3c01c
Bug-Url: https://bugzilla.redhat.com/1346270
Signed-off-by: Arik Hadas <ahadas@redhat.com>