webadmin: Warn for running VMs when cluster level change
When changing cluster level of a cluster with running VMs,
a warning is shown stating requirement to reboot all these VMs once
the cluster level is changed.

All running/paused non-[Hosted Engine|External] VMs are marked with
the NextRun configuration mark to denote the requirement.
The HE VM is handled in other path.

Prior this fix, the cluster level change was not allowed
when there were running VMs.

This fix solves issue with hosted engine VM running at the time of
cluster level change.

Change-Id: I8628d6794248076599087bd22aa3aca44fde69e2
Bug-Url: https://bugzilla.redhat.com/1348907
Signed-off-by: Marek Libra <mlibra@redhat.com>