restapi: Fix adding of quota storage limits
The quota limit resources use inheritance and generics in order to avoid
implementing the "add" operation twice, once for cluster limits and
another time for storage limits. Unfortunately the JAX-RS implementation
that we use doesn't recognize these generic "add" operations, resulting
in a 405 error message when they are used. To avoid that error this
patch refactors the relevant resources so that they don't use
inheritance and generics. The result is more verbose, but simpler, and
works.

Change-Id: I4ee930ecbf6bd6d477f3746eaa409082688cd5a4
Bug-Url: https://bugzilla.redhat.com/1348559
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>