restapi: Don't try to set IPs for VMs without id
Currently the V3 adapter for VM objects tries to populate the IP
addresses reported by the guest agent, but this doesn't work for VM
objects that haven't been created yet, like the ones returned by the
operation to import a VM from an export storage domain.

In addition retrieving the details of a VM before importing it doesn't
work either, because the implementation uses inheritance and generics
and the JAX-RS implementation that we use in version 4 of the engine
doesn't understand it.

This patch fixes these two isses needed to make the import operation
succeed in both versions of the API.

Change-Id: I2e2ea61f34467bcbc0578fde7ba8a0fa9d12a648
Bug-Url: https://bugzilla.redhat.com/1348805
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>