restapi: Map correctly disk interface types
Currently the mapping of disk interfaces is implemented using string
translation, thus assuming that the names of the enums are exactly the
same in the backend and in the API. But that isn't true. This patch
changes that mapping so that it uses type safe enums.

Change-Id: I0fabd55154eb78de271cbf3805375181e144d2ea
Bug-Url: https://bugzilla.redhat.com/1349594
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 84ff17784ea1784df62bd6daf42b53c1ee3e422e)