restapi: Always populate V3 disk interface and bootable
Currently the V3 disk interface and bootable attributes are only
populated with a disk is retrieved, but it isn't populated for disk
objects returned by the "add" or "update" operations. To fix that
this patch moves the code that populates these attributes from the
implementation of the "get" operation to the adapter that transforms
all output disk objects.

Change-Id: I17634e05969ca826ae9a1c7aa08d9bfcfe0ef620
Bug-Url: https://bugzilla.redhat.com/1349594
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>