restapi: Don't generate streaming support code
When version 4 of the API was planned it was decided to introduce
support for streaming results. This required the introduction of
streaming XML and JSON parsing and rendering code, automatically
generated from the model. But this code isn't currently used, and
generating increases the compilation time and incorrectly reduces the
percentage of coverage reported by some tools. To avoid those
inconveniences this patch removes the generation of this code from the
bulid process. This will be reintroduced later, when the streaming
support is completed.

Change-Id: Iff09f64dc755493facaa4c9b7e457860c4982f18
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>