core: Take VM lock before assigning a pooled VM to user
In AttachUserToVmFromPoolAndRunCommand, take an exclusive lock on the VM
before setting user's ownership on it. This prevents other commands to
act on this VM before it is run as the result of
AttachUserToVmFromPoolAndRunCommand. On another side, if the VM is
already locked by some other command (e.g. RunVmCommand),
AttachUserToVmFromPoolAndRunCommand will not be able to lock it and will
be forced to take another VM from the pool.

Change-Id: Id2290f7b229041b989011f2acd4bf96af51357f9
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1350403
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>