webadmin: Unneeded Vdsm call when editing iSCSI SD
When editing an existing iSCSI, we need to get from Vdsm the list of
available LUNs.

However, due to multiple events triggered, we are sending the VDSM query
more than needed.

In this patch, we are minimizing the handling of the events triggered and
therefore the queries from 6 to 2 in the edit SD flow.

Change-Id: Ic6f5db4f74efaad05a6b84e14d12a700efc27b50
Bug-Url: https://bugzilla.redhat.com/1259353
Signed-off-by: Fred Rolland <frolland@redhat.com>