core: fix PowerSavingWeight score calculation
EvenDistributionWeightPolicy unit calculated a weight score that was
comprised from CPU usage and memory consumption.
This calculation was wrong (which Power saving weight module relied on)
there was a need to separate the weight policy units to 2 units :
EvenDistributionCPUWeightPolicy and EvenDistributionMemoryWeightPolicy
which produced 2 new power saving units:
PowerSavingMemoryWeightPolicyUnit and PowerSavingMemoryCPUPolicyUnit
instead of PowerSavingWeightPolicyUnit.

CPU usage is calculated as before
(power saving unit takes the opposite score)
Memory score is calculated by
the range of 1 to the maximum available memory of a host in the cluster
and normalized to the score range.

NonWeightPolicy unit was removed and will replaced
with memory and CPU even distribution policies under
the default 'None' scheduling policy

Change-Id: I4ed2fa107b4c1139a4a4f9c07894472280acbf6b
Bug-Url: https://bugzilla.redhat.com/1260381
Signed-off-by: Yanir Quinn <yquinn@redhat.com>
Signed-off-by: yquinn <yquinn@redhat.com>
(cherry picked from commit 6037d18)