restapi: Return added affinity label
Currently when an affinity label is added to a host or to a virtual
machine, the doesn't return the added label, like all the other "add"
operations of the API do. This patch changes the relevant resource so
that it will fetch and return the added affinity label.

Change-Id: I8d8c9de6b87424706faf7643924790ba08200c4b
Bug-Url: https://bugzilla.redhat.com/1350228
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 57b26e6e4abd81389937fe931a0cfdb5af879cca)