core: evict event from cache when manually removed
When a event or alert is removed manually from the display in UI, we
should also evict the corresponding cache entry for events that have 6he
flood mechanism.
This patch evict the corresponding event from the cache when it is
manually deleted by the user.
This solves the following problem :

Event E1 with flood time of T1 is raised by the system
The issue is fixed by the application administrator
Event E1 is raised again within the T1 period

The bug was that the second event was suppressed by the flooding
mechanism since the cache still holds the first event.

In order to solve this BZ there was a need to add code to the
RemoveAuditLogByIdCommand to evict the event from the cache, this is
done by composing the cache key from the command using the information
from the event which is the candidate for the delete operation and
generating a this key with the sysadmin user.

Change-Id: I0e302aa0bb218f7986cac57b13a90f689d956e19
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1219147
Signed-off-by: emesika <emesika@redhat.com>