Revert "webadmin: Add label comboboxes to VM & host dialogs"
The bug (https://bugzilla.redhat.com/show_bug.cgi?id=1338799) is targeted to
4.1, should not be part to 4.0.

Also the feature does not make much sense in edit dialogs since it is a read
only property. Also the placement of it inside the dialog is questionable. The
implementation itself is also not really aligned with the rest of the code.

This reverts commit 61d899dd1b19ebd475ef3187ca74c9cf1360da35.

Bug-Url: https://bugzilla.redhat.com/1338799
Change-Id: I4e13cd4651a614d68d7dd7462094f9180c7479d6
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>