core: Generate JBoss module for extensions API directly
Currently the JBoss module for the extensions API is generated as part
of the "dependencies" module. As a result the "dependencies" module
can't be built independent of the rest of the project, because the
extensions API artifact isn't available unless the extensions API module
is part of the build. This patch changes the build process so that the
JBoss module is built as part of the build of the extensions API module
itself, so it doesn't need to be included in the "dependencies" module.

Change-Id: Idf4e93f7629094d2f82b127c4a4b26613faa87cc
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>