gluster: support FQDN for server address in brick
Currently only IP address is supported for gluster bricks
with gluster network defined. This fails to sync bricks
which are created in CLI using FQDN of the gluster network.

 This patch address this issue by resolving the FQDN to
correct IP address and compare that address with the
actual IP address of nic with gluster network

Note: This patch will work only if ENGINE can resolve the host address
used in the brick to correct IP address in the host.

Change-Id: I0c596dcaa119b20135bb56d15ff5f6672875bb41
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>