core: Pass storage pool ID to disk images on storage validation
Commit 2c6791 introduced a refactoring of whether images exist on storage
check, however it used the storage pool ID taken from the images where
in the case of import the image storage pool ID contains the original
storage pool the images resided on which may not be the storage pool we
import the image to.
The storage pool ID was restored from the original behavior and is now
taken from the parameters of the commands.

Change-Id: I505a70ccb142030d5db140c9da7b03eab939f646
Signed-off-by: Tal Nisan <tnisan@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1352422