restapi: Convert job and step status to upper case
In version 4 of the API all the enums are represented using lower case,
including the job and steps status. The values are by default copied
without modification by the version 3 adapters. But some enums used to
be represented using upper case in version 3 of the API, in particular
job and step status used to be upper case, so we need to convert them.

Change-Id: I237ce19f5995ce27647650bdf085227ae89e7d2d
Bug-Url: https://bugzilla.redhat.com/1352575
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>