engine: Teardown image after verifying upload.
When uploading an image finishes, the uploading session is revoked and
the image is being verified in VDSM. currently in the revoking function
(stopImageTransferSession) there's a call to teardown image, which makes
it unreachable to VDSM when trying to verify it.

Move the teardown call for after stopping the session. this way only when
we finish using the image it will be teared down.

If the stopImageTransferSession fails, don't tear down the image, as the
session still alive, thus someone might still using it.

Change-Id: I7b6587452c2d2879dd97ed2645dd4c348cb0175b
Bug-Url: https://bugzilla.redhat.com/1352676
Signed-off-by: Amit Aviram <aaviram@redhat.com>