core: Remove local DC for shared SD
A local DC can now contain shared SDs and local SDs.
Removing a local DC used to format the master domain
attached to it since local SDs can not be detached.

Shared SDs behave differently, since those can be detached.
Based on that, if a local DC will be removed the master storage domain
will be formatted only if it is a local SD (to keep the previous
behaviour) and if it is a shared SD then this storage domain will only
be detached.

In the future we should have a seperate patch which will let the user
choose whether to format the last local SD or only detach and
remove it so it can be imported later.

Change-Id: I2385ef4a1a8d36761a0359fe420722f7b4a3bf90
Bug-Url: https://bugzilla.redhat.com/1302185
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>