core: refactor VmInternalData
Replace all constructors with one constructor that receives the notification
time. All other arguments should be set with setter methods. This simplifies
the instanciation of this class.

In addition, statically import the methods in VdsBrokerObjectsBuilder that
are used in VmStatsVdsBrokerCommand and in ListVDSCommand in order to make
it more readable.

Change-Id: Iceeacc97412fdba7bb54d1d1c4139d800a45528a
Signed-off-by: Arik Hadas <ahadas@redhat.com>