core: Add a shared domains comparator when reconstruct.
When electing a new master, reconstruction is attempted according to an
ascending order of the last time a domain was used as a master.

In a DC that contains both shared and local domains,
we should add a secondary criteria to prefer shared domains over local
domains for resiliency considerations.

Change-Id: Iaaa15e56531c046a45eaf201b0d989dbfc7e8517
Bug-Url: https://bugzilla.redhat.com/1353134
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>