core: Cluster update triggers VM update only if necessary
There was a problem when user tried to edit Cluster while there was a
migrating VM in the Cluster. The root cause was collision of VM locks
of MigrateVmCommand and UpdateVmCommand called by UpdateClusterCommand.

This patch doesn't solve the problem entirely. It only lowers the
probability its occurrence by triggering UpdateVmCommand only if cluster
compatibility level was changed.

Change-Id: Ieccac1f08847681d5f326b9c54a99364720cfdb0
Bug-Url: https://bugzilla.redhat.com/1349526
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>