core: persist before validate() only when there is a parent with cb
Currently we persist the command entity before the validate() check and on
failure the command status is updated to ENDED_WITH_FAILURE.
We do that so that When there is a parent command with a callback, it will be
able to detect that the child command execution failed on that phase. Additionally,
the parent is responsible to clear the command hirerchy.
When there is no parent command, the command entity will never be cleared/monitored.

In this patch the logic is changed so that the command is persisted before the CDA
phase only in case there's a parent with a CB, otherwise it'll be persisted on the
execution phase.

Change-Id: Iae96cb414677c970c91ef860824d2a5ca981d861
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1352825
Signed-off-by: Liron Aravot <laravot@redhat.com>