webadmin: v2v-improve loading of external vms
Improve the loading performance of external VMs from external server
for the following sources: VMware, KVM, Xen.

This is done by receiving the VMs names only in the 1st phase, i.e. in
the 1st import dialog (New VDSM api was added for receiving
VMS names only on a seperate patch).
And only when choosing the VMS to import and clicking on the "Next"
button in order to display full VMs data on the 2nd dialog,
then we receive full info (xml+disks) for those choosen VMs only.
(changing the existed VDSM api for receiving full
VMs data on subset of VMS and not for all VMs existed was added in a
seperate patch).

The following functional changes were done to support it:
1. Displaying only VMs names in those two lists of 1st dialog. The CPU
architecture for each VM which was displayed in brackets was removed.

2. Checking the cpu architecture of the chosen VMS in order to find
appropriate cluster is done only after clicking the "Next" button
because we don't have this information earlier. In case of a problem
with the cpu architecture, a popup message is displayed as before.

3. The VMs status (up/down) is known only after clicking the "Next"
button. Therefore in case the user choose VMs in 1st dialog which are
discovered to be running or can't be retrieved with their full info,
they won't appear on 2nd dialog, BUT a warning message in
the appropriate dialog is displayed in such cases and
also a message in audit log listing the running VMs is
added.
So the user will be able to stop/fix those VMs, click "Back" and then
"Next" and those missing VMs will appear in the 2nd dialog.

4. Each clicking on the "Next" button "reloads" all VMs data for
all chosen VMs and not just for the added new ones.

5. backward compatibility -  for handling old proxy hosts:
When choosing old proxy host < 4.1 then the flow is the old flow as
done before this fix.
Meaning that we get all VMs with full data in the 1st dialog.
The only changes for old proxy hosts are that no cpu architecture is
displayed and also better and new messages and logging in case of
running/non-retrieved VMs choosen to be imported.

6. Adding this new patch support for 4.1 and upper versions is
configurable via the vdc_options and can be disabled if needed.

Bug-Url: https://bugzilla.redhat.com/1294629
Change-Id: Ica95a150c656ad9b860914323b05ba466837be77
Signed-off-by: Sharon Gratch <sgratch@redhat.com>