core: remove inaccessible memory during upgrade
This patch fixes an issue discovered in the upgrade process where the
upgrade fails while trying to convert memory volumes on a storage
domain that does not exist in the system to disk images.

So before doing the conversion, such memory volumes are first being
removed. They are unusable anyway.

Change-Id: Iead5deab3c572f6bbc02af1f56c41ffdfdf0fc82
Bug-Url: https://bugzilla.redhat.com/1353219
Signed-off-by: Arik Hadas <ahadas@redhat.com>