restapi: Fix validation of pool when adding permission
When adding a permission we check that there is an object provided, a
data center, a virtual machine, etc. But the validation for the pool
uses an incorrect name, "vmpool" instead of "vmPool". As a result the
validation always fail, and it is impossible to add a permission for a
pool. This patch fixes that issue.

Change-Id: I8be122dc6be608f5df85864be50d8faa0032806c
Bug-Url: https://bugzilla.redhat.com/1354463
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit a6782813d66fecec7ff2f32e610375e4ac6527f9)