core: prevent floating vms in unknown status
On non-responsive treatment, update VMs to unknown status only
if they run on some host. This check is done in the database
level, with the protection of the database lock. This way, if
the monitoring has just updated the VM to down state we will
not end up with a VM with unknown status and run_on_vds = NULL.

This is not the ideal solution but it seems to cover the most
problematic scenario and yet simple.

Change-Id: I48c205e9228a32822148332af13c89e51cc80afa
Bug-Url: https://bugzilla.redhat.com/1354494
Signed-off-by: Arik Hadas <ahadas@redhat.com>