core: CommandBase - don't update the command entity always
In change Iae96c executeAction() was changed to only update the command
entity if present after the execution.
On flows with async tasks a CommandEntity is being created
for the root command with special values during the execution,
the update done in CommandBase will override those values to different ones.

This patch sets the update to be done only on commands with cb/parent command
with cb to avoid updating the unneeded entities.

Change-Id: I68af8d469ce9f42c8b1230fc4bef23448345b59a
Bug-Url: https://bugzilla.redhat.com/1352825
Signed-off-by: Liron Aravot <laravot@redhat.com>