core: StorageDomainNumberOfVmDisksComparator streaming
The StorageDomainNumberOfVmDisksComparator constructor converts a
collections of DiskImages to a map from the storage domain id to the
number of VMs located on it.

This is much easier to do with Java 8's steam APIs instead of
reinventing the wheel ourselves.

As a side effect of this patch, the Map was converted from
Map<Guid, Integer> to Map<Guid, Long> to conform to
Collectors.counting()'s API. This may add some minor memory
consumption overhead, as Longs are larger than Integers, but this is
negligible as this Comparator is extremely short lived.

Change-Id: Id8271fd39eb270090f25ebd9bd32351d61fa465b
Signed-off-by: Allon Mureinik <amureini@redhat.com>