backend: Filter orphaned Host Devieces on refresh
In some hotplug scenarios (especially those related to SCSI devices) it can
happen that libvirt reports to VDSM an internally inconsistent list of devices,
namely that the parent_device points to a device not in the list or is empty
altogether.

Those cases violate the DB constraint imposed on the host_device and cause the
RefreshHostDevicesCommand transaction to fail. Since those cases are impossible
to handle at lower layers and we need to maintain consistend view of the host
devices for our main business logic, we now ignore those invalid devices in the
hostdev refresh cycle.

Change-Id: I520d7b6cc7ef099da7b6a67381a7cdc8d2ae0e64
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1315100
Signed-off-by: Martin Betak <mbetak@redhat.com>