core: vms monitoring no longer always expects vm statistics
VmStatistics are processed and persisted only on statistics monitoring
cycles. However, while analyzing the VM we used to count on having the
statistics also for regular (non-statistics) monitoring cycles. This
required as to have a workaround in the handling of events, which do
not include the statistics (the statistics were loaded from the DB).

Now that we use VmStatistics data only on statistics monitoring cycles
we can remove this workaround.

Change-Id: If5d69f975b4245f0ae05ba4ccfec1fa24a57237f
Signed-off-by: Arik Hadas <ahadas@redhat.com>