core: Remove StorageDomainFilter
StorageDomainFilter was a wrapper for two methods - an abstract method
returning a Predicate and a filterStorageDomains method that applied
that filtered a List<StorageDomain> by that predicate. Since in Java 8's
filtering a list is so easy, there's really no reason to have this in a
dedicated class. This patch removes the StorageDomainFilter class and
just keeps the Predicates and moves the filtering to
MemoryStorageHandler.

A side effect of this patch is a slightly improved performance of the
filtering. Even though the current implementation potentially still has to
perform M (filters) * N (domains) evaluations, it collects the result
only once, instead of having to construct a list M times.

Change-Id: I3795e2f81919fbb6ed92636343d04ed913de937f
Signed-off-by: Allon Mureinik <amureini@redhat.com>