core: MemoryStorageHandler simple predicates
With Java 8's predicates, The StorageDomainStatusFilter and
StorageDomainTypeFilter become one-line lambdas expression, and it's a
bit pointless to maintain them as their own classes. This patch
removes them and replaces them with public static final members in
MemoryStorageHandler.

As their respective tests were just re-implementing the exact same
logic and checking it against itself. This kind of test is somewhat
pointless, so this patch just removes them.

Change-Id: I49849a5cc1d97875448462b8e65891d2451976c2
Signed-off-by: Allon Mureinik <amureini@redhat.com>