core: show descriptive error on illegal max value
Show descriptive error on illegal max value in search engine.

This patch fixes an issue when illegal negative max value is passed
to the search engine.

The problem was that the default value for the max parameter was -1 which
was translated as "no max parameter given" so all records should be
retrieved.
Instead of that, the default value was changed to Integer.MAX_VALUE
In addition, in case of RuntimeException, the exception was swallowed by
SearchQuery::initQueryData method and therefor the error was not
returned to the API call.
I changed that such that a RuntimeException will be thrown again after
writing to the log.

Change-Id: I1821b4a8109da4c48c8d8d32c0c52dadd1603b78
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1352953
Signed-off-by: emesika <emesika@redhat.com>