core: BaseImagesCommand - isDataOperationsByHSM() logic
SDM flows would be implemented in such a way that in case that the host
which is performing the operation will go non responsive we'll be able
to poll using a different host by check the actual entity on the storage.
In order to do that on oVirt images, we intend to use the volume leases
so we can determine if the operation on the volume was ended or is still
running.
Export domains and local storage domains doesn't support volume leases
therefore if the performing host goes non responsive there's no way to track
the operation, therefore the SDM verbs aren't supposed to support those.

This patches update isDataOperationsByHSM() to return false if the domain
isn't data domain or if it is a local domain.

Change-Id: I7826ef8317a667d2a563a702a5e6931f382f638d
Signed-off-by: Liron Aravot <laravot@redhat.com>