core: Refactoring of VM from pool allocation process
All functionality of looking for an available VM in a pool and taking
locks on it is moved to a separate class VmPoolHandler. Both
VmPoolMonitor and AttachUserToVmFromPoolAndRunCommand are using it.
Every VM that needs to be run is locked at the moment it is found. This
approach helps to prevent race conditions and makes any internal lock
objects redundant.

Change-Id: I6bbc6d650e0a706f7ff9970ac2dd425e20b6bfc8
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>