core: extract vm numa updates from resource manager
It is wrong to clear VM numa nodes in DestroyVmVdsCommand because at
that point we cannot be sure that the VM was really destroyed. We
should have waited for the monitoring to detect that the VM is indeed
down and only then clear the VM numa nodes.

But that is the only place that currently rely on internalSetVmStatus
in ResourceManager to clear VM numa nodes. We want this method to
receive only VmDynamic eventually, so as part of this patch I just
extract the updates of VM numa nodes from it to the destroy command.

Change-Id: I504829db92044f4c0a1df41b9f14e2304ce6608f
Signed-off-by: Arik Hadas <ahadas@redhat.com>